GuardianCare_BackEnd/services/asr-service/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
filelock==3.17.0
Flask==3.1.0
fsspec==2025.3.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
PyJWT==2.10.1
numba==0.61.0
numpy==2.1.3
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
regex==2024.11.6
requests==2.32.3
setuptools==76.0.0
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
